完成商品添加功能 --图片添加功能待完成

+++ taotao-manager/taotao-manager-interface/src/main/java/com/taotao/manager/service/IBaseService.java
@@ -18,7 +18,7 @@ public interface IBaseService<T> {
 
     public void update(T t);
 
-    public T selectById(int id);
+    public T selectById(Long id);
 
     public List<T> selectAll();
